requirements.txt
aioredis==1.3.1
asgiref==2.3.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.0
channels==2.4.0
channels-redis==2.3.2
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
daphne==2.5.0
Django==2.2.14
django-cors-headers==3.4.0
django-filter==2.3.0
django-redis==4.12.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msgpack==0.5.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pytz==2020.1
redis==3.5.3
requests==2.22.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.9
zope.interface==5.1.0
